working database (boardgames), working confing and parsing

+++ app/src/main/java/models/BoardGame.kt
@@ -6,6 +6,6 @@ package models
         var yearPublished: Int,
         var image: String,
         var thumbnail: String,
-        var bggId: Long
+        var bggId: Int
     )
